Move the experimental API into jupedsim.h
